+++ Formula/netductor.rb
@@ -1,26 +1,26 @@
 class Netductor < Formula
   desc "Netductor control plane CLI / TUI"
   homepage "https://github.com/PavelNeyman/netductor"
-  version "0.8.82"
+  version "0.8.83"
   license "MIT"
   on_macos do
     on_arm do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.82/netductor-darwin-arm64"
-      sha256 "c49b85f450888ffc15109675426104cb639913f7b59a993e515f524c30e9f06f"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.83/netductor-darwin-arm64"
+      sha256 "0fe6991413f2a69d28ef70fcd714b77e92166cd8b49c35dcf3e6d3d2936cc4b6"
     end
     on_intel do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.82/netductor-darwin-amd64"
-      sha256 "3210bf01ef798998461f46a3c2673e6f93440608316736df56581bdbf044bef1"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.83/netductor-darwin-amd64"
+      sha256 "b15b2b99fa7f1de463dab40129a9702f5adb04a8e64fabf0119c217ceffc2832"
     end
   end
   on_linux do
     on_intel do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.82/netductor-linux-amd64"
-      sha256 "3cb2194f0f8392c35b4a7348eb112d64511268ed8c978974a345934c4a7626e6"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.83/netductor-linux-amd64"
+      sha256 "dfa8ac689435e5a319a12b7e63a7d122680962665aa05e53b48c41eac9fc3277"
     end
     on_arm do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.82/netductor-linux-arm64"
-      sha256 "ea234220acb3c0788382c73c7261a0922c83f3119ffbda74f40a8d63c8d3731f"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.83/netductor-linux-arm64"
+      sha256 "a56dd0a2afdd17cb1f315448330db6679452f5ea00637b387d58e2d7088efca7"
     end
   end
   def install